requirements.txt
python-telegram-bot==21.5
python-dotenv==1.0.1
aiosqlite==0.20.0
python-dateutil==2.9.0
qrcode[pil]==7.4.2
Pillow==10.4.0
requests==2.32.3